feat: new capability interface / new UI number display

+++ src/main/java/cn/howxu/mmcr/internal/port/EnergyHatchSize.java
@@ -6,20 +6,20 @@ package cn.howxu.mmcr.internal.port;
  * @author howxu <dev@howxu.cn>
  */
 public enum EnergyHatchSize {
-    TINY("tiny", 2048, 128),
-    SMALL("small", 4096, 512),
-    NORMAL("normal", 8192, 512),
-    REINFORCED("reinforced", 16384, 2048),
-    BIG("big", 32768, 8192),
-    HUGE("huge", 131072, 32768),
-    LUDICROUS("ludicrous", 524288, 131072),
-    ULTIMATE("ultimate", 2097152, 131072);
+    TINY("tiny", 2048L, 128L),
+    SMALL("small", 4096L, 512L),
+    NORMAL("normal", 8192L, 512L),
+    REINFORCED("reinforced", 16384L, 2048L),
+    BIG("big", 32768L, 8192L),
+    HUGE("huge", 131072L, 32768L),
+    LUDICROUS("ludicrous", 524288L, 131072L),
+    ULTIMATE("ultimate", 2097152L, 131072L);
 
     private final String id;
-    private final int capacity;
-    private final int transfer;
+    private final long capacity;
+    private final long transfer;
 
-    EnergyHatchSize(String id, int capacity, int transfer) {
+    EnergyHatchSize(String id, long capacity, long transfer) {
         this.id = id;
         this.capacity = capacity;
         this.transfer = transfer;
@@ -29,11 +29,11 @@ public enum EnergyHatchSize {
         return id;
     }
 
-    public int capacity() {
+    public long capacity() {
         return capacity;
     }
 
-    public int transfer() {
+    public long transfer() {
         return transfer;
     }
 }
